File analysis
Assignment 03 Details

Your program is able to decode/display the message in Challenge3-1.bmp, Challenge3-2.bmp, and Challenge3-3.bmp.

Your program will allow the user to set the bit depth of the output .bmp file.

Your program will allow the user to select from red, green, blue, yellow, white, or black as the colour in which to display the decoded message.

Your program will read and display the files in the current directory and allow the user to select the file to read from and decode. (Google: "python directory listing" for this. Alternatively "python select file dialog" if you wanted to use tkinter to open a file dialog box and return a file name selected by the user.)

I do not expect formal headers as in other programs. I do expect significant inline commenting that will allow me to understand what your code is doing or you to understand it after not having looked at it for a year.

Nice to have: An option to display all header info (as in Challenge 1)