Banking
Assignment 01 Details

Write a banking application that a teller in a bank might use to record in-person customer transactions for multiple customers.

The program will have a minimum menu interface as follows.
Set customer – sets the user name for the current customer so the program knows which customer to deposit to or withdraw from. (N.B. needs to be done before other menu items can be called.)
Deposit – records a deposit to the customer’s chequing or savings account
Withdraw – records a withdrawal from the customer’s chequing or savings account
N.B. the amount to withdraw cannot exceed the balance for the account
Balance – displays the balance in both accounts for the current customer
Exit – exits the program

The menu loops repeatedly until the user chooses the exit option and the program takes the appropriate action based on the menu choice.

Customer user names, first and last names, and balances in their chequing and savings accounts are stored in a text file called customers.txt. Data fields in customers.txt are separated by commas (csv == comma separated values). For example:
User name          First name          Last name          Chequing balance          Savings balance
jdoe                     Jane                    Doe                    123.45                              678.90

Your program must read data from customers.txt and, when appropriate, write data to it. Customer information needs to persist from one run of the program to another.

Please refer to ICS4U_0201_funcDocumentation.py for an example of the documentation expectations for this assignment.

Implementing the above requirements with reasonable consideration to the user interface will earn you a level 4- mark for this assignment.
Structure of the main program
I recommend that you use the following logic to implement your menu and main program. The idea is that the user makes a menu choice and the if...elif executes a function which matches the menu choice. Once the function exits, program flow returns to the while loop and the menu is displayed again ready for the next action.

WHILE not pgm_exit
PUT “(1) Set customer”
PUT “(2) Deposit”
PUT “(3) Withdraw”
PUT “(4) Balance”
PUT “(5) Exit”

PROMPT for menu choice (u_menu)
IF u_menu == “1”
	CALL set customer function
ELIF u_menu == “2”
	CALL deposit function
ELIF u_menu == “3”
	CALL withdraw function
ELIF u_menu == “4”
	CALL balance function
ELIF u_menu == “5”
	SET pgm_exit = TRUE
ENDIF
ENDWHILE

Note that after each function call in the if…elif we return to the top of the loop and present the menu to the user again. This is the main body of our program. Almost all the rest of the code should be in functions that you write.

The set customer function
The set customer function is particularly important and you should write it first. You need a set customer function so that the program knows which customer will be deposited to or withdrawn from. The set customer function will need to perform the following tasks:
Prompt for the customer's user name
Open the customers.txt file
Read through the file and see if the customer's user name exists in the file (See Reading/Writing customers.txt below)
If the customer's user name is found, set a variable to hold it and return it to the main program.
If the customer's user name is NOT found, print an error message and return None to the main program

Reading/Writing customers.txt
You should write a utility function that opens the customers.txt file in read mode and loops through the file line by line. Split each line into a list (using the comma as the separator) and append it to a list of user account information. It’s much easier to make any changes to the particular user’s data when it's in this list and then write the data for all users back to the customers.txt file when you’re ready. (Since you already have all the data for the users in the list in your program, you don’t have to worry about overwriting the data in customers.txt when you open it for writing.)

As an alternative, you could just append the lines from customers.txt to a user account list as strings (i.e. not split the lines themselves into a list at this point). This would mean that later on, whenever you needed to access a particular field in the list (e.g. the customer's chequing balance), you'd need to split the string into its individual fields.

You should also write a complementary utility function that opens customers.txt in write mode and writes out the contents of your user account information list to customers.txt in the proper format.

I recommend that you use strip() on any strings that you read in to your program from the customers.txt file in order to remove any new line characters at the end of the lines. This will avoid some issues that can arise and make sure you are only dealing with the strings of characters you are interested in. It will also remove any blank lines that you might have added to the customers.txt file by mistake that might otherwise cause problems.
Checking for the existence of the customers.txt file
It’s a good idea to check that the file exists before you try to read from or write to it. Instead of using try…except, you can import python’s os package and use a built-in command to simplify this. For example:

import os

#blah, blah, blah

if not(os.path.exists(file_name)):
print("ERROR: %s does not exist. Cannot proceed" % file_name)
else:
	#write your code here

Enhancements
You may implement any enhancements you wish as long as you meet the basic menu requirements as outlined above. The following are some suggestions.
Paying special/extra attention to clarity and aesthetic appeal of the user interface.
Add functionality that allows the user to add a new customer. New customer information will have to be stored in the customers.txt file.
Add functionality that allows money to be transferred from the customer’s chequing account to their savings account and vice versa.
Add functionality that records all transactions and the date and time on which they occurred (use python’s date-time functionality for this) and displays them to the user.
A further enhancement would allow the user to display the transaction history for a particular customer.