Summative
Summative Details

All invoice information is contained in a file called Level4InvoiceInformation.txt.
It can be found in the handout folder in the
Summative subdirectory.
The file contains “Header” records and “Detail” records.
Header records are preceded by a line in the data file that is comprised of a single entry -:
the letter “H”. Header records contain customer information (name, address, etc.). A
Header record contains the following information in this specific order.
1. Account Number
2. Customer Name
3. Customer Street Address
4. Customer City
5. Customer Province
6. Customer Postal Code
7. Invoice Date
8. Invoice Number
9. Shipping method
a. PC – Priority Courier – $35.00
b. XP – Express Post – $20.00
c. RP – Regular Post – $10.00

Detail records are preceded by a line in the data file that is comprised of a single entry:
the letter “D” and contain information about particular purchases. Each Detail record
contains the following information in this specific order:
1. Item Number
2. Item Description
3. Quantity Sold
4. Unit Price
Your program must read the information from the text file and produce an invoice. Note
that there is information for multiple invoices in the file. Your program must process all
invoices.
Tax on the invoice must be calculated according to the province of the purchaser. The tax
calculation will differ from province to province. (See Appendix B – Calculating
Provincial Sales Tax.)
Your program will display a formatted invoice on the screen and write the invoice to an
output file. A sample invoice layout is provided in Appendix A – Sample Invoice.