Intro to Dictionaries Part B
Intro to Dictionaries Part B Details

Each entry in the list is a string starting with an airport code and followed by some country information. 
The airport code and the country info are separated by a colon (":")
Using the airport data list that has been posted:
loop thru the list and split each element at the colon
treat the airport code as the key and the country info as the value
add the key-value pairs to a dictionary
allow the user to loop and enter an airport code.
your program will return the city that matches the airport code
Does your code handle an invalid airport code?